server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
sqlalchemy[asyncio]>=2.0.35
aiosqlite>=0.20.0
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sentence-transformers>=3.1.0
numpy>=1.26.0
python-ulid>=2.7.0
httpx>=0.27.0
